fix: make shortcuts work as expected
Makes shortcuts such as ctrl-1 and Super-1 work instead of being
recognized by windows and applicatios as ctrl-5 (the original number in
the standard layout).
This is done by changing the scan codes for the keys in the number row.
This may cause some issues in some apps such as vim.